refactored POC and added metadata
