--- .idea/.gitignore
@@ -1,4 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
-/.idea*
