+++ models/models.go
@@ -2,16 +2,19 @@ package models
 
 import (
 	"fmt"
+	"strings"
 	"time"
 )
 
 // ImportRow is a Go representation of an area profiles key statistic in an import cvs row.
 type ImportRow struct {
-	AreaCode string
-	Title    string
-	Name     string
-	Value    string
-	Unit     string
+	AreaCode    string
+	Title       string
+	Name        string
+	Value       string
+	Unit        string
+	DatasetID   string
+	DatasetName string
 }
 
 // Area is a simplified domain representation of a Geographical area. Other fields omitted for the sake of this POC.
@@ -38,14 +41,23 @@ type AreaProfile struct {
 
 // KeyStatistic is a domain model representing a key statistical figure for an area profile.
 type KeyStatistic struct {
-	VersionID    int       `json:"version_id,,omitempty"`
-	StatID       int       `json:"id"`
-	ProfileID    int       `json:"profile_id"`
-	Name         string    `json:"name"`
-	Value        string    `json:"value"`
-	Unit         string    `json:"unit"`
-	DateCreated  time.Time `json:"date_created"`
-	LastModified time.Time `json:"last_modified,omitempty"`
+	VersionID    int                  `json:"version_id,,omitempty"`
+	StatID       int                  `json:"id"`
+	ProfileID    int                  `json:"profile_id"`
+	Name         string               `json:"name"`
+	Value        string               `json:"value"`
+	Unit         string               `json:"unit"`
+	DateCreated  time.Time            `json:"date_created"`
+	LastModified time.Time            `json:"last_modified,omitempty"`
+	Metadata     KeyStatisticMetadata `json:"metadata,omitempty"`
+}
+
+// KeyStatisticMetadata is a domain model representing metadata associated with a KeyStatistic
+type KeyStatisticMetadata struct {
+	MetadataID  int    `json:"id"`
+	DatasetID   string `json:"dataset_id"`
+	DatasetName string `json:"dataset_name"`
+	Link        string `json:"href"`
 }
 
 // KeyStatsVersion is model containing verison metda data about a an area profile key statistic.
@@ -62,4 +74,8 @@ type KeyStatsVersion struct {
 func (k KeyStatistic) ToString() string {
 	format := "[StatID: %d, ProfileID: %d, Name: %s, Value: %s, Unit: %s, Date Created: %+v]"
 	return fmt.Sprintf(format, k.StatID, k.ProfileID, k.Name, k.Value, k.Unit, k.DateCreated)
+}
+
+func (r ImportRow) GetDatasetHref() string {
+	return fmt.Sprintf("http://localhost:666/datasets/%s/%s", r.DatasetID, strings.ToLower(strings.Replace(r.DatasetName, " ", "_", -1)))
 }
