+++ testdata/ex1.csv
@@ -1,7 +1,7 @@
-AreaCode,Title,Name,Value,Unit
-E05011362,"Resident Population for Disbury East, Census 2021",Resident population,"503,127",
-E05011362,"Resident Population for Disbury East, Census 2021",Population density (Hectares),"43.5",
-E05011362,"Resident Population for Disbury East, Census 2021",Average (mean) age,"29",
-E05011362,"Resident Population for Disbury East, Census 2021",People think their general health is good,"31.7",%
-E05011362,"Resident Population for Disbury East, Census 2021",Households where English is not the main language,"10.3",%
-E05011362,"Resident Population for Disbury East, Census 2021","Households owned with a mortgage, loan or shared ownership","22.6",%
+AreaCode,Title,Name,Value,Unit,Dataset ID,Dataset Name
+E05011362,"Resident Population for Disbury East, Census 2021",Resident population,"503,127",,abc123,Test dataset 1
+E05011362,"Resident Population for Disbury East, Census 2021",Population density (Hectares),"43.5",,efg789,Test dataset 2
+E05011362,"Resident Population for Disbury East, Census 2021",Average (mean) age,"29",,abc123,Test dataset 1
+E05011362,"Resident Population for Disbury East, Census 2021",People think their general health is good,"31.7",%,efg789,Test dataset 2
+E05011362,"Resident Population for Disbury East, Census 2021",Households where English is not the main language,"10.3",%,abc123,Test dataset 1
+E05011362,"Resident Population for Disbury East, Census 2021","Households owned with a mortgage, loan or shared ownership","22.6",%,xxx666,Test dataset 3
